search_app/requirements.txt
google-cloud-logging
google-cloud-storage
flask==1.1.2
pymysql
pyyaml==5.3.1
SQLAlchemy==1.3.17